Change Hazel as a static library and fix some warnings.

+++ Hazel/src/Hazel/Input.h
@@ -10,7 +10,7 @@ namespace Hazel {
 
 		inline static bool IsMouseButtonPressed(int button) { return s_Instance->IsMouseButtonPressedImpl(button); }
 		
-		inline static std::pair<float, float> GetMousePosition() { return s_Instance->GetMousePositionImpl(); }
+		inline static std::pair<double, double> GetMousePosition() { return s_Instance->GetMousePositionImpl(); }
 
 		inline static float GetMouseX() { return s_Instance->GetMouseXImpl(); }
 		
@@ -20,7 +20,7 @@ namespace Hazel {
 	protected:
 		virtual bool IsKeyPressedImpl(int keycode) = 0;
 		virtual bool IsMouseButtonPressedImpl(int button) = 0;
-		virtual std::pair<float, float> GetMousePositionImpl() = 0;
+		virtual std::pair<double, double> GetMousePositionImpl() = 0;
 		virtual float GetMouseXImpl() = 0;
 		virtual float GetMouseYImpl() = 0;
